Title: 앞에서부터 삭제하기 2, Time: 816ms, Memory: 34MB, Status: Passed - Codetree

+++ 250515/앞에서부터 삭제하기 2/delete-it-from-the-beginning-2.java
@@ -23,7 +23,7 @@ public class Main {
             sum += arr[n-i];
             size++;
 
-            ans = Math.max(ans, (sum - q.peek()) / (size - 1));
+            ans = Math.max(ans, (double)(sum - q.peek()) / (double)(size - 1));
         }
 
         System.out.format("%.2f\n", ans);
